+++ intelmq/bots/experts/taxonomy/expert.py
@@ -29,7 +29,7 @@ TAXONOMY = {
     "unauthorised-information-modification": "information-content-security",
     "data-loss": "information-content-security",
     "dropzone": "information-content-security",  # not in ENISA eCSIRT-II taxonomy
-    "leak": "information-content-security",  # not in ENISA eCSIRT-II taxonomy
+    "data-leak": "information-content-security",
     "scanner": "information-gathering",
     "sniffing": "information-gathering",
     "social-engineering": "information-gathering",